requirements.txt
flask
flask-cors
google-generativeai>=0.7.2
PyPDF2
redis